requirements.txt
# Core Dependencies
google-generativeai>=0.3.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0

# Google API Dependencies
google-api-python-client>=2.0.0,<3.0.0
google-auth-oauthlib>=1.0.0,<2.0.0
google-auth-httplib2>=0.1.0,<1.0.0
google-auth>=2.0.0,<3.0.0

# Database
sqlite3>=3.35.0  # Python standard library, version for documentation
